[FIX] 로그인 엑티비티 수정, 비밀번호 초기화 로직 변경, 로그인 뷰모델 수정 UserModel petGender 토글버튼으로 수정 -> Int 값으로 변경
