+++ app/src/main/java/com/example/frompet/ui/chat/activity/ChatUserDetailActivity.kt
@@ -76,7 +76,7 @@ class ChatUserDetailActivity : AppCompatActivity() {
         binding.apply {
             tvPetName.text = user.petName
             tvPetAge.text = "${user.petAge.toString()}세"
-            tvPetGender.text = user.petGender
+            tvPetGender.text = user.petGender.toString()
             tvPetType.text = user.petType
             tvPetDes.text = user.petDescription
             tvPetIntro.text = user.petIntroduction
