+++ app/src/main/java/com/example/frompet/ui/login/LoginupViewModel.kt
@@ -1,5 +1,6 @@
 package com.example.frompet.ui.login
 
+import android.content.ContentValues.TAG
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -38,97 +39,31 @@ class LoginupViewModel @Inject constructor(
     }
 
     fun signInUser(email: String, password: String) = viewModelScope.launch {
-        val currentUser = auth.currentUser
-        Log.d("zzzzzzzz", "Current User: ${auth.currentUser}")
-        if (currentUser != null) {
-            // 이미 로그인한 사용자가 있습니다. 중복 로그인을 방지합니다.
-            viewModelScope.launch(Dispatchers.Main) {
-                eventsChannel.send(AllEvents.Error("이미 로그인한 사용자가 있습니다."))
+        when {
+            email.isEmpty() -> {
+                eventsChannel.send(AllEvents.ErrorCode(1))
             }
-        } else {
-            // 로그인 시도
-            when {
-                email.isEmpty() -> {
-                    eventsChannel.send(AllEvents.ErrorCode(1))
-                }
-                password.isEmpty() -> {
-                    eventsChannel.send(AllEvents.ErrorCode(2))
-                }
-                else -> {
-                    auth.signInWithEmailAndPassword(email, password)
-                        .addOnCompleteListener { task ->
-                            if (task.isSuccessful) {
-                                val user = task.result?.user
-                                if (user != null) {
-                                    _firebaseUser.postValue(user)
-                                    _loginResult.postValue(true)
-                                } else {
-                                    _firebaseUser.postValue(null)
-                                    _loginResult.postValue(false)
-                                }
-                            } else {
-                                val exception = task.exception
-                                if (exception is FirebaseAuthException) {
-                                    if (exception.errorCode == "INVALID_LOGIN_CREDENTIALS") {
-                                        viewModelScope.launch(Dispatchers.Main) {
-                                            eventsChannel.send(AllEvents.Error("이메일 또는 비밀번호가 올바르지 않습니다."))
-                                        }
-                                    } else {
-                                        viewModelScope.launch(Dispatchers.Main) {
-                                            eventsChannel.send(AllEvents.Error(exception.message ?: "로그인 중 오류 발생"))
-                                        }
-                                    }
-                                } else {
-                                    viewModelScope.launch(Dispatchers.Main) {
-                                        eventsChannel.send(AllEvents.Error("로그인 중 오류 발생"))
-                                    }
-                                }
-                            }
-                        }
-                }
+            password.isEmpty() -> {
+                eventsChannel.send(AllEvents.ErrorCode(2))
+            }
+            else -> {
+                actualSignInUser(email , password)
             }
         }
     }
-
-     /*       else -> {
-                auth.signInWithEmailAndPassword(email, password)
-                    .addOnCompleteListener { task ->
-                        if (task.isSuccessful) {
-                            val user = task.result?.user
-                            if (user != null) {
-                                _firebaseUser.postValue(user)
-                                _loginResult.postValue(true)
-                            } else {
-                                _firebaseUser.postValue(null)
-                                _loginResult.postValue(false)
-                            }
-                        } else {
-                            // Firebase Authentication 로그인 실패
-                            val exception = task.exception
-                            if (exception is FirebaseAuthException) {
-                                if (exception.errorCode == "INVALID_LOGIN_CREDENTIALS") {
-                                    // 이메일 또는 비밀번호가 올바르지 않음
-                                    viewModelScope.launch(Dispatchers.Main) {
-                                        eventsChannel.send(AllEvents.Error("이메일 또는 비밀번호가 올바르지 않습니다."))
-                                    }
-                                } else {
-                                    // 다른 Firebase Authentication 오류 처리
-                                    viewModelScope.launch(Dispatchers.Main) {
-                                        eventsChannel.send(AllEvents.Error(exception.message ?: "로그인 중 오류 발생"))
-                                    }
-                                }
-                            } else {
-                                // Firebase Authentication 예외 발생
-                                viewModelScope.launch(Dispatchers.Main) {
-                                    eventsChannel.send(AllEvents.Error("로그인 중 오류 발생"))
-                                }
-                            }
-                        }
-                    }
+    private fun actualSignInUser(email:String, password: String) = viewModelScope.launch {
+        try {
+            val user = userRepository.signInWithEmailPassword(email, password)
+            user?.let {
+                _firebaseUser.postValue(it)
+                eventsChannel.send(AllEvents.Message("로그인 성공"))
             }
+        }catch(e:Exception){
+            val error = e.toString().split(":").toTypedArray()
+            Log.d(TAG, "signInUser: ${error[1]}")
+            eventsChannel.send(AllEvents.Error(error[1]))
         }
-    }*/
-
+    }
     fun signUpUser(email: String, password: String, confirmPass: String) = viewModelScope.launch {
         when {
             email.isEmpty() -> {
@@ -145,12 +80,22 @@ class LoginupViewModel @Inject constructor(
             }
 
             else -> {
-                val user = userRepository.signUpWithEmailPassword(email, password)
-                if (user != null) {
-                    _firebaseUser.postValue(user)
-                    _loginResult.postValue(false)
+                actualSignUpUser(email, password)
                 }
             }
+        }
+
+    private fun actualSignUpUser(email:String , password: String) = viewModelScope.launch {
+        try {
+            val user = userRepository.signUpWithEmailPassword(email, password)
+            user?.let {
+                _firebaseUser.postValue(it)
+                eventsChannel.send(AllEvents.Message("회원가입 성공"))
+            }
+        }catch(e:Exception){
+            val error = e.toString().split(":").toTypedArray()
+            Log.d(TAG, "signInUser: ${error[1]}")
+            eventsChannel.send(AllEvents.Error(error[1]))
         }
     }
 
@@ -202,6 +147,5 @@ class LoginupViewModel @Inject constructor(
         data class Message(val message: String) : AllEvents()
         data class ErrorCode(val code: Int) : AllEvents()
         data class Error(val error: String) : AllEvents()
-        data class PasswordCheck(val message: String) : AllEvents()
     }
 }
