--- app/src/main/java/com/example/frompet/ui/login/UserViewModel.kt
@@ -1,146 +1,0 @@
-package com.example.frompet.ui.login
-
-import android.util.Log
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.example.frompet.data.repository.user.BaseAuthRepository
-import com.google.firebase.auth.FirebaseUser
-import com.google.firebase.firestore.FirebaseFirestore
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.receiveAsFlow
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-@HiltViewModel
-class UserViewModel @Inject constructor(
-    private val repository: BaseAuthRepository,
-    private val firestore: FirebaseFirestore
-):ViewModel() {
-    private val TAG = "UserViewModel"
-
-    private val _firebaseUser = MutableLiveData<FirebaseUser?>()
-    val currentUser get() =_firebaseUser
-
-    private val eventsChannel = Channel<AllEvents>()
-
-    val allEventsFlow = eventsChannel.receiveAsFlow()
-
-
-    fun signInUser(email: String , password: String) = viewModelScope.launch{
-        when {
-            email.isEmpty() -> {
-                eventsChannel.send(AllEvents.ErrorCode(1))
-            }
-            password.isEmpty() -> {
-                eventsChannel.send(AllEvents.ErrorCode(2))
-            }
-            else -> {
-                actualSignInUser(email , password)
-            }
-        }
-    }
-
-    fun signUpUser(email : String , password: String , confirmPass : String)= viewModelScope.launch {
-        when{
-            email.isEmpty() -> {
-                eventsChannel.send(AllEvents.ErrorCode(1))
-            }
-            password.isEmpty() -> {
-                eventsChannel.send(AllEvents.ErrorCode(2))
-            }
-            password != confirmPass ->{
-                eventsChannel.send(AllEvents.ErrorCode(3))
-            }
-            else -> {
-                actualSignUpUser(email, password)
-            }
-        }
-    }
-
-
-    private fun actualSignInUser(email:String, password: String) = viewModelScope.launch {
-        try {
-            val user = repository.signInWithEmailPassword(email, password)
-            user?.let {
-                _firebaseUser.postValue(it)
-                eventsChannel.send(AllEvents.Message("로그인 성공"))
-            }
-        }catch(e:Exception){
-            val error = e.toString().split(":").toTypedArray()
-            Log.d(TAG, "signInUser: ${error[1]}")
-            eventsChannel.send(AllEvents.Error(error[1]))
-        }
-    }
-
-    private fun actualSignUpUser(email:String , password: String) = viewModelScope.launch {
-        try {
-            val user = repository.signUpWithEmailPassword(email, password)
-            user?.let {
-                _firebaseUser.postValue(it)
-                eventsChannel.send(AllEvents.Message("회원가입 성공"))
-            }
-        }catch(e:Exception){
-            val error = e.toString().split(":").toTypedArray()
-            Log.d(TAG, "signInUser: ${error[1]}")
-            eventsChannel.send(AllEvents.Error(error[1]))
-        }
-    }
-
-    fun signOut() = viewModelScope.launch {
-        try {
-            val user = repository.signOut()
-            user?.let {
-                eventsChannel.send(AllEvents.Message("로그아웃 실패"))
-            }?: eventsChannel.send(AllEvents.Message("로그아웃 성공"))
-
-            getCurrentUser()
-
-        }catch(e:Exception){
-            val error = e.toString().split(":").toTypedArray()
-            Log.d(TAG, "signInUser: ${error[1]}")
-            eventsChannel.send(AllEvents.Error(error[1]))
-        }
-    }
-
-    fun getCurrentUser() = viewModelScope.launch {
-        val user = repository.getCurrentUser()
-        _firebaseUser.postValue(user)
-    }
-
-    fun verifySendPasswordReset(email: String){
-        if(email.isEmpty()){
-            viewModelScope.launch {
-                eventsChannel.send(AllEvents.ErrorCode(1))
-            }
-        }else{
-            sendPasswordResetEmail(email)
-        }
-
-    }
-
-    private fun sendPasswordResetEmail(email: String) = viewModelScope.launch {
-        try {
-            val result = repository.sendResetPassword(email)
-            if (result){
-                eventsChannel.send(AllEvents.Message("이메일로 전송되었습니다"))
-            }else{
-                eventsChannel.send(AllEvents.Error("could not send password reset"))
-            }
-        }catch (e : Exception){
-            val error = e.toString().split(":").toTypedArray()
-            Log.d(TAG, "signInUser: ${error[1]}")
-            eventsChannel.send(AllEvents.Error(error[1]))
-        }
-    }
-
-
-    sealed class AllEvents {
-        data class Message(val message : String) : AllEvents()
-        data class ErrorCode(val code : Int): AllEvents()
-        data class Error(val error : String) : AllEvents()
-    }
-}
-
-
