+++ app/src/main/res/values/themes.xml
@@ -7,4 +7,8 @@
     </style>
 
     <style name="Theme.FromPet" parent="Base.Theme.FromPet" />
+
+    <style name="CalenderViewDateCustomText" parent="android:TextAppearance.DeviceDefault.Small">
+        <item name="android:textColor">@android:color/black</item>
+    </style>
 </resources>
